requirements.txt
openai
httpx
requests
python-dotenv
